requirements.txt
# === Web scraping & download ===
niquests>=3.6.0          # Drop-in replacement for requests with native HTTP/2
beautifulsoup4>=4.12.0
lxml>=4.9.0

# === PDF text extraction ===
pdfplumber>=0.10.0
PyPDF2>=3.0.0


# === OCR for scanned PDFs ===
# Using EasyOCR (GPU/MPS acceleration if available)
easyocr>=1.7.1
torch>=2.0.0
Pillow>=10.0.0
# pytesseract>=0.3.10  # (deprecated, replaced by EasyOCR)

# === Data manipulation ===
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0
openpyxl>=3.1.0

# === Visualization (Sprint 2+) ===
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# === NLP (Sprint 2+) ===
spacy>=3.7.0
# Spanish model (run once after install):
#   python -m spacy download es_core_news_lg
sentence-transformers>=2.2.0
bertopic>=0.16.0
hdbscan>=0.8.33
scikit-learn>=1.3.0

# === Graphs (Sprint 2+) ===
networkx>=3.1
# Gephi is a desktop app, not a Python library:
#   https://gephi.org

# === LLM (Sprint 1+) ===
anthropic>=0.25.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.13.0
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
bleach==6.3.0
bokeh==3.8.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
colorcet==3.1.0
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.6
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
easyocr==1.7.2
et_xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.25.2
fsspec==2026.3.0
h11==0.16.0
holoviews==1.22.1
httpcore==1.0.9
httpx==0.28.1
hvplot==0.12.2
idna==3.11
ImageIO==2.37.3
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jh2==5.0.10
Jinja2==3.1.6
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
lazy-loader==0.5
linkify-it-py==2.1.0
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdit-py-plugins==0.5.0
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.18.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
ninja==1.13.0
niquests==3.18.2
numpy==2.4.3
openai==2.30.0
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
panel==1.8.9
param==2.3.2
parso==0.8.6
pdfminer.six==20251230
pdfplumber==0.11.9
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
plotly==6.6.0
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyPDF2==3.0.1
pypdfium2==5.6.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyviz_comms==3.0.6
PyYAML==6.0.3
pyzmq==27.1.0
qh3==1.7.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-image==0.26.0
scipy==1.17.1
setuptools==81.0.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
stack-data==0.6.3
sympy==1.14.0
tifffile==2026.3.3
torch==2.11.0
torchvision==0.26.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uc-micro-py==2.0.0
urllib3==2.6.3
urllib3-future==2.19.900
wassima==2.0.5
wcwidth==0.6.0
webencodings==0.5.1
xyzservices==2025.11.0
